chore(gate, sdk): upload artifacts before deploying typegraph
