+++ typegate/src/engine/query_engine.ts
@@ -114,13 +114,6 @@ export interface EndpointToSchemaMap {
   [index: string]: { fnName: string; outputSchema: unknown };
 }
 
-export interface UploadUrlMeta {
-  fileName: string;
-  fileHash: string;
-  fileSizeInBytes: number;
-  urlUsed: boolean;
-}
-
 export class QueryEngine {
   name: string;
   queryCache: QueryCache;
@@ -138,7 +131,6 @@ export class QueryEngine {
       }
     >
   >;
-  fileUploadUrlCache: Map<string, UploadUrlMeta> = new Map();
 
   get rawName(): string {
     return this.tg.rawName;
